moving grab pass to graphics folder too
